updated TiddlySpaceInit to restore avatar if necessary
It is possible for space members to delete their SiteIcon, which leaves
a rather ugly red box whenever that space's avatar is to be used (e.g.
in the backstage).

Thus if no avatar exists in the current space's bags, the default is
restored.

NB: The heuristic employed here can be fooled by a private SiteIcon
existing without a public equivalent. That was deemed acceptable.